refactor: deduplicate P-256 field arithmetic into embedded-cal::p256
