+++ embedded-cal/src/lib.rs
@@ -3,6 +3,7 @@
 #![no_std]
 
 pub mod empty;
+pub mod p256;
 
 mod aead;
 mod dh;
